+++ crates/common/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.1.1-git
+## 0.1.1
 
 ### Minor
 
@@ -10,4 +10,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 2 -->
+<!-- Increment to skip CHANGELOG.md test: 0 -->